# Improve performance of installing gems from gem server sources