--- a/bundler/lib/bundler/cli.rb
+++ b/bundler/lib/bundler/cli.rb
@@ -251,9 +251,7 @@ module Bundler
       remembered_negative_flag_deprecation("no-deployment")
 
       require_relative "cli/install"
-      Bundler.settings.temporary(:no_install => false) do
-        Install.new(options.dup).run
-      end
+      Install.new(options.dup).run
     end
 
     map aliases_for("install")
@@ -299,9 +297,7 @@ module Bundler
     def update(*gems)
       SharedHelpers.major_deprecation(2, "The `--force` option has been renamed to `--redownload`") if ARGV.include?("--force")
       require_relative "cli/update"
-      Bundler.settings.temporary(:no_install => false) do
-        Update.new(options, gems).run
-      end
+      Update.new(options, gems).run
     end
 
     desc "show GEM [OPTIONS]", "Shows all gems that are part of the bundle, or the path to a given gem"