--- a/bundler/lib/bundler/rubygems_integration.rb
+++ b/bundler/lib/bundler/rubygems_integration.rb
@@ -206,16 +206,10 @@ module Bundler
     def spec_from_gem(path, policy = nil)
       require "psych"
       gem_from_path(path, security_policies[policy]).spec
-    rescue Exception, Gem::Exception, Gem::Security::Exception => e # rubocop:disable Lint/RescueException
-      if e.is_a?(Gem::Security::Exception) ||
-          e.message =~ /unknown trust policy|unsigned gem/i ||
-          e.message =~ /couldn't verify (meta)?data signature/i
-        raise SecurityError,
-          "The gem #{File.basename(path, ".gem")} can't be installed because " \
-          "the security policy didn't allow it, with the message: #{e.message}"
-      else
-        raise e
-      end
+    rescue Gem::Security::Exception => e
+      raise SecurityError,
+        "The gem #{File.basename(path, ".gem")} can't be installed because " \
+        "the security policy didn't allow it, with the message: #{e.message}"
     end
 
     def build_gem(gem_dir, spec)