--- a/bundler/lib/bundler/rubygems_integration.rb
+++ b/bundler/lib/bundler/rubygems_integration.rb
@@ -205,10 +205,6 @@ module Bundler
 
     def spec_from_gem(path, policy = nil)
       gem_from_path(path, security_policies[policy]).spec
-    rescue Gem::Security::Exception => e
-      raise SecurityError,
-        "The gem #{File.basename(path, ".gem")} can't be installed because " \
-        "the security policy didn't allow it, with the message: #{e.message}"
     end
 
     def build_gem(gem_dir, spec)