## What was the end-user or developer problem that led to this PR?

This originally started because I wanted to do a fix involving printing a warning when certain gemspecs are unpacked from the downloaded `.gem` package and I noticed that for every gem installed, Bundler creates two `Gem::Package` instances and unpacks the specification from each of them. So the warning would be printed twice.

## What is your fix for the problem, implemented in this PR?

My fix is to make sure a single `Gem::Package` is created when installing each gem.

This turns out to provide speed and memory improvements 🎉.

For example using dependency files from the activeadmin repository at https://github.com/activeadmin/activeadmin/pull/7543 (modified to remove the `sassy-rails` dependency since otherwise it eats up most `bundle install` time, and configuring bundler to install to `vendor/bundle`), I get the following results:

```
✗ hyperfine 'BUNDLER_VERSION=2.4.0.dev bundle install' 'BUNDLER_VERSION=2.3.15 bundle install' --prepare 'rm -rf vendor/bundle'
Benchmark 1: BUNDLER_VERSION=2.4.0.dev bundle install
  Time (mean ± σ):      8.377 s ±  0.267 s    [User: 11.841 s, System: 6.485 s]
  Range (min … max):    8.017 s …  8.869 s    10 runs
 
Benchmark 2: BUNDLER_VERSION=2.3.15 bundle install
  Time (mean ± σ):      9.146 s ±  0.149 s    [User: 13.060 s, System: 7.185 s]
  Range (min … max):    8.910 s …  9.404 s    10 runs
 
Summary
  'BUNDLER_VERSION=2.4.0.dev bundle install' ran
    1.09 ± 0.04 times faster than 'BUNDLER_VERSION=2.3.15 bundle install'
```

And with memory

```
➜  activeadmin git:(add-darwin) ✗ rm -rf vendor/bundle && BUNDLER_VERSION=2.3.15 ruby-memory-profiler --pretty --no-detailed --allocated-strings=0 --retained-strings=0 /Users/deivid/.asdf/installs/ruby/3.1.2/lib/ruby/gems/3.1.0/gems/bundler-2.3.15/exe/bundle install -- --quiet
Total allocated: 1.06 GB (5656647 objects)
Total retained:  88.13 MB (1187607 objects)

➜  activeadmin git:(add-darwin) ✗ rm -rf vendor/bundle && BUNDLER_VERSION=2.4.0.dev ruby-memory-profiler --pretty --no-detailed --allocated-strings=0 --retained-strings=0 /Users/deivid/.asdf/installs/ruby/3.1.2/lib/ruby/gems/3.1.0/gems/bundler-2.4.0.dev/exe/bundle install -- --quiet
Total allocated: 918.79 MB (5373031 objects)
Total retained:  88.10 MB (1188919 objects)
```

## Make sure the following tasks are checked

- [x] Describe the problem / feature
- [ ] Write [tests](https://github.com/rubygems/rubygems/blob/master/bundler/doc/development/PULL_REQUESTS.md#tests) for features and bug fixes
- [x] Write code to solve the problem
- [x] Make sure you follow the [current code style](https://github.com/rubygems/rubygems/blob/master/bundler/doc/development/PULL_REQUESTS.md#code-formatting) and [write meaningful commit messages without tags](https://github.com/rubygems/rubygems/blob/master/bundler/doc/development/PULL_REQUESTS.md#commit-messages)