@Wolfer 2022-06-21
`bundle package --no-install` doesn't work yet. How set this option now?