@deivid-rodriguez 2022-06-21
Oh, sorry, I just saw what you mean. It seems like a regression, I'll fix it.