@deivid-rodriguez 2022-06-21
https://github.com/rubygems/rubygems/pull/5639 will fix this, thanks for letting me know!